requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
MarkupSafe==3.0.2
numpy==2.2.2
pandas==2.2.3
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
requests